e2e_cascading/requirements.txt
torch
torchaudio
transformers
librosa
soundfile
scikit-learn
PyYAML

